requirements.txt
pandas>=2.0.0
openpyxl>=3.1.0
python-Levenshtein>=0.21.0  # For fuzzy string matching in PHI detection
faker>=19.0.0  # For generating realistic fake data
